feat: add runtime type selector
A Runtime type selector has been added, allowing users to switch between different Runtime types. In this way, we can use the standard url and format attributes of the DataTable component, and at the same time view all types of Runtime data.Although this implementation approach is not the initially envisioned "displaying all Runtime types simultaneously", it is a reasonable compromise solution for the current component limitations. If a true multi-runtime type merge display is to be achieved in the future, it is necessary to consider modifying the back-end API or creating a custom DataTable component.